captcha-training/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
onnxruntime>=1.17.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
albumentations>=1.3.1
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
